+++ src/Feature/Metadata/MetadataCompletionKind.php
@@ -5,6 +5,7 @@ namespace Symfony\Lsp\Feature\Metadata;
 enum MetadataCompletionKind
 {
     case FormOption;
+    case FormProperty;
     case Constraint;
     case ConstraintOption;
     case SerializerGroup;